* speech-dispatcher.texi (Speech Synthesis Commands in C): Formatting.
* src/tests/Makefile.am (bin_PROGRAMS): connection_recovery
-- new target.

* src/c/api/libspeechd.h (SPD_FATAL): Flush stdout.

* src/c/api/libspeechd.c (spd_close): Join th events thread.
(spd_say): Check reply.
(spd_say): Set msg_id to -1 on error.
(spd_execute_command): Logging message.
(spd_execute_command_wo_mutex): Logging message.
(spd_execute_command_with_reply): Return -1 on error.
(spd_send_data): Return NULL on error in spd_send_data_wo_mutex.
(spd_send_data): Return NULL if connection->stream is NULL.
(spd_send_data_wo_mutex): Only print out the reply if it is not NULL.
(spd_events_handler): Formatting.
(spd_events_handler): Signal reply ready even if socket is broken.
(xmalloc): Formatting.

+++ src/modules/module_utils.c
@@ -18,7 +18,7 @@
  * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA.
  *
- * $Id: module_utils.c,v 1.50 2007-11-26 17:07:18 hanke Exp $
+ * $Id: module_utils.c,v 1.51 2008-02-08 10:01:09 hanke Exp $
  */
 
 #include "fdsetconv.h"
@@ -481,7 +481,7 @@ size_t
 module_parent_wfork(TModuleDoublePipe dpipe, const char* message, EMessageType msgtype,
                     const size_t maxlen, const char* dividers, int *pause_requested)
 {
-    size_t pos = 0;
+    unsigned int pos = 0;
     char msg[16];
     char *buf;
     int bytes;
